Revert "Rename Integration module to MonkeyPatching". Fixes #94.
It turns out that all but the HEAD version of `ActiveSupport` are using
`Mocha::Integration::TestUnit::AssertionCounter` even though it is an
undocumented class.

I only renamed the module in the first place to make room for a better
(non-monkey-patching) test frameworkk integration solution, but given
this will be breaking things for a lot of people, it seems best to
revert the change.