Move Test::Unit monkey-patching code into the individual version modules.
By doing this I hope to make it easier to do something different for the
most recent versions where we want to use the new adapter.

This commit does introduce significant duplication, but I hope to remove
that later. And encouragingly the duplicate code is now almost identical
to the MiniTest equivalent - see
7ed2733b7da21d62e34ae6bf6209861877adb76b.