Only run "adapter" tests for different versions of MiniTest & Test::Unit
* The so-called "adapter" tests are really integration tests in the
sense that they check that Mocha is properly integrated with the
relevant test library.

* Running *all* the tests for all the versions of MiniTest & Test::Unit
was overkill. It's only really the integration between Mocha and the
test library that is likely to differ in these cases, so the new
"adapter" tests are sufficient.

* I think I've made the same changes to both the build-matrix script and
and to .travis.yml - it's clearly a pain to have to maintain both of
these, but fixing that is a job for another day.

* We're still not running tests for the 3 earlier versions of MiniTest,
because they don't work for some reason. I had thought it was something
to do with running all the tests, but now it looks like it might just be
that the TestRunner doesn't work with these versions of MiniTest in
Ruby 1.9.3.

* In running these tests, a minor bug in most of the monkey-patch versions
of Test::Unit integration came to light and these are fixed in this
commit. This is gives me some confidence that there is a decent amount
of coverage in the "adapter" tests.