Make unit tests more robust to changes in environment. Fixes #121.
@voxik wanted to run the tests by simply requiring the files rather than
running the rake task. If the unit tests and acceptance tests are loaded
in the same process and if a recent version of MiniTest is available,
then `ExpectationErrorFactory.exception_class` was being set to
`MiniTest::Assertion` at file load time and then later when the unit
tests ran, the latter exception was being raised instead of the default
`ExpectationError` exception. This causes a number of unit tests to
fail.

I have fixed this by changing the assertions to expect
`ExpectationErrorFactory.exception_class`. I've also had to change
`ExpectationErrorFactory.exception_class` to be set to the default at
file load time to make this work.