Allow stubbing of private Kernel methods. Fixes #134.
We were only testing stubbing of the public and protected methods on
Object.

Note that `#method_defined?` and `#instance_methods` include both public
and protected, but not private methods.

There are still a few methods that are awkward, but I've just
excluded them from the tests for now.

There's probably a better solution to this in 1.9 involving inheriting
Mock from BasicObject instead of Object, but that's a job for another
day.