Make auto-activation of Test::Unit integration more resilient.
This change is specifically to cope with the nasty re-defining of
classes that is done by the `minitest-spec-rails` gem [1]. This
confuses the Test::Unit version detection logic in Mocha and it thinks
that the Ruby 1.8 standard library version of Test::Unit is being used
when in fact there is no Test::Unit version present.

I'm not completely sure why this has reared its head [2] with Rails 4
and/or Ruby 2, and haven't checked carefully that this change is
(a) completely logical; and (b) hasn't introduced any regressions,
but it does seem to fix the problem. I'm going to push it now and try to
track down any regressions separately - I'm pretty hopeful there won't
be any.

Note that it will be possible to avoid this problem in the future
when we introduce the ability to manually activate specific test library
integration and stop using the automatic activation.

[1] https://github.com/metaskills/minitest-spec-rails/issues/17
[2] https://github.com/freerange/mocha/issues/143