Define Mocha module before it's referenced.
The test helper defines a class `TestCase` within the `Mocha` module. When
running the tests inside the bundle, the `Mocha` module happens to be defined
at this point. However when running the tests outside the bundle, it is not
defined and so an exception is raised:

    uninitialized constant Mocha (NameError)

This commit should fix this problem and hopefully it fixes #163.