Do not automatically try to integrate with test libraries.
Since the automatic test library integration functionality requires the test
library to be loaded and this doesn't usually happen until *after* the bundle
is loaded, it makes things simpler if we use `require 'mocha/setup'` to
explicitly setup Mocha when we know the test library has been loaded.

I deprecated the use of `require 'mocha'` in v0.13.0 and I plan to release
this change as part of a major version bump (i.e. v1.0.0), so it ought to
be safe to remove the deprecation in `lib/mocha.rb`.