Reset environment variables between builds.
I was seeing a problem where `MOCHA_RUN_INTEGRATION_TESTS` was being set in one
build but not set in a subsequent build. The value from the first build was
trumping the lack of one in the subsequent one.

It might be simpler to prefix commands with the environment variables to avoid
so much "global state", but this will do for now.